Exit with status 1 on error
Add macros `with-error-status' and `with-error-and-help' that help to
easily report errors to the user. The latter also calls the `help'
command.

All commands, when done, exit with `*exit-status*' being the status
code used. The `with-error-status' macro let-binds this variable so
that the command function exits with the given status code.